e2e-demo/pyproject.toml
[project]
name = "edge-operator-agent"
version = "0.1.0"
description = "Edge Operator Agent for industrial/manufacturing scenarios using Strands Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[ollama]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.0",
    "streamlit>=1.30.0",
    "hypothesis>=6.0.0",
    "mcp>=1.0.0",
    "PyPDF2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
